db_saver/requirements.txt
opensearch-py
pymilvus
kafka-python
requests
python-dotenv
uuid
